# docs/openapi.json
{
  "components": {
    "schemas": {
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ValidationDetail"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Details"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "PaginationResponse": {
        "properties": {
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "per_page": {
            "title": "Per Page",
            "type": "integer"
          },
          "total_items": {
            "title": "Total Items",
            "type": "integer"
          },
          "total_pages": {
            "title": "Total Pages",
            "type": "integer"
          }
        },
        "required": [
          "page",
          "per_page",
          "total_items",
          "total_pages"
        ],
        "title": "PaginationResponse",
        "type": "object"
      },
...
      "SportResponse": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "max_players": {
            "title": "Max Players",
            "type": "integer"
          },
          "max_players_in_game": {
            "title": "Max Players In Game",
            "type": "integer"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "max_players",
          "max_players_in_game"
        ],
        "title": "SportResponse",
        "type": "object"
      },
...
      "TeamCreateRequest": {
        "additionalProperties": false,
        "properties": {
          "gender_category": {
            "enum": [
              "male",
              "female"
            ],
            "title": "Gender Category",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sport_id": {
            "exclusiveMinimum": 0.0,
            "title": "Sport Id",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "sport_id",
          "gender_category"
        ],
        "title": "TeamCreateRequest",
        "type": "object"
      },
      "TeamEnvelope": {
        "properties": {
          "team": {
            "$ref": "#/components/schemas/TeamResponse"
          }
        },
        "required": [
          "team"
        ],
        "title": "TeamEnvelope",
        "type": "object"
      },
      "TeamListResponse": {
        "properties": {
          "pagination": {
            "$ref": "#/components/schemas/PaginationResponse"
          },
          "teams": {
            "items": {
              "$ref": "#/components/schemas/TeamResponse"
            },
            "title": "Teams",
            "type": "array"
          }
        },
        "required": [
          "teams",
          "pagination"
        ],
        "title": "TeamListResponse",
        "type": "object"
      },
      "TeamResponse": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "disabled_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disabled At"
          },
          "gender_category": {
            "enum": [
              "male",
              "female"
            ],
            "title": "Gender Category",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_enabled": {
            "title": "Is Enabled",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sport": {
            "$ref": "#/components/schemas/SportResponse"
          }
        },
        "required": [
          "id",
          "name",
          "sport",
          "gender_category",
          "is_enabled",
          "created_at",
          "disabled_at"
        ],
        "title": "TeamResponse",
        "type": "object"
      },
      "TeamUpdateRequest": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "TeamUpdateRequest",
        "type": "object"
      },
...
      "ValidationDetail": {
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message",
          "type"
        ],
        "title": "ValidationDetail",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "AccessTokenAuth": {
        "bearerFormat": "JWT",
        "description": "JWT access token. It expires after 15 minutes.",
        "scheme": "bearer",
        "type": "http"
      },
      "RefreshTokenAuth": {
        "bearerFormat": "JWT",
        "description": "Rotating JWT refresh token. Use it only for refresh or logout.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "summary": "Authentication, Sports, and Teams management API.",
    "title": "Sports App API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/teams": {
      "get": {
        "description": "Returns a filtered and paginated Team catalogue. Enabled Teams are returned by default and each page contains at most 25 records.",
        "operationId": "teamsList",
        "parameters": [
          {
            "description": "Partial accent-insensitive Team-name search.",
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "description": "Partial accent-insensitive Team-name search.",
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "sport_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "exclusiveMinimum": 0.0,
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Sport Id"
            }
          },
          {
            "in": "query",
            "name": "gender_category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "male",
                    "female"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": null,
              "title": "Gender Category"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "enabled",
              "enum": [
                "enabled",
                "disabled",
                "all"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "name_asc",
              "enum": [
                "name_asc",
                "created_at_desc"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "exclusiveMinimum": 0.0,
              "title": "Page",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamListResponse"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "List Teams",
        "tags": [
          "Teams"
        ]
      },
      "post": {
        "description": "Creates an enabled Team. Its identity is unique by normalized name, Sport, and gender category.",
        "operationId": "teamsCreate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamEnvelope"
                }
              }
            },
            "description": "Created"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Create a Team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{team_id}": {
      "get": {
        "description": "Returns one enabled or disabled Team by identifier.",
        "operationId": "teamsGet",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Get a Team",
        "tags": [
          "Teams"
        ]
      },
      "put": {
        "description": "Changes only the name of an enabled Team. Sport and gender are immutable.",
        "operationId": "teamsUpdate",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TeamUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Rename a Team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{team_id}/disable": {
      "patch": {
        "description": "Disables a Team without deleting it. Repeated requests are idempotent.",
        "operationId": "teamsDisable",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Disable a Team",
        "tags": [
          "Teams"
        ]
      }
    },
    "/teams/{team_id}/enable": {
      "patch": {
        "description": "Enables a Team again. Repeated requests are idempotent.",
        "operationId": "teamsEnable",
        "parameters": [
          {
            "in": "path",
            "name": "team_id",
            "required": true,
            "schema": {
              "title": "Team Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamEnvelope"
                }
              }
            },
            "description": "OK"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "security": [
          {
            "AccessTokenAuth": []
          }
        ],
        "summary": "Enable a Team",
        "tags": [
          "Teams"
        ]
      }
    }
  },
...
}